validation logic moved from the FindBookForm component to the FormDataAndValidation file
